Fullscreen mode with SDL
Full screen mode can be toggled with ALT-Enter. The overall experience
is quite nice, but the entire content of the application is shown in the
top right corner of the screen.

To finish this feature, I fear support from PDCurses is required to draw
the contents centered or even scale everything up to to use the
available space.

Closes #285